requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pandas
python-dotenv
